Normalize Sage Terminal module layout
