+++ app/terminal/src/backend/runtime/mod.rs
@@ -1,0 +1,11 @@
+mod cli;
+mod root;
+mod state;
+
+#[cfg(test)]
+pub(crate) use cli::resolve_python_command;
+pub(crate) use cli::{resolve_cli_invoker, run_cli_json_owned, CliInvoker};
+pub(crate) use root::resolve_runtime_root;
+#[cfg(test)]
+pub(crate) use root::resolve_runtime_root_from;
+pub(crate) use state::{apply_state_env, prepare_state_root};
