+++ app/terminal/src/backend/tests/contract.rs
@@ -1,0 +1,79 @@
+use serde_json::json;
+
+use crate::backend::contract::{expect_array_field, parse_stream_event, CliJsonCommand};
+use crate::backend::protocol::parse_backend_line;
+use crate::backend::BackendEvent;
+
+#[test]
+fn parse_stream_event_collects_tool_fields_from_multiple_locations() {
+    let event = parse_stream_event(
+        r#"{
+            "type":"tool_call",
+            "tool_calls":[{"function":{"name":"write_file"}}],
+            "metadata":{"tool_name":"shell"},
+            "tool_name":"exec"
+        }"#,
+    )
+    .expect("stream event should parse");
+
+    let parsed = parse_backend_line(
+        r#"{
+            "type":"tool_call",
+            "tool_calls":[{"function":{"name":"write_file"}}],
+            "metadata":{"tool_name":"shell"},
+            "tool_name":"exec",
+            "content":"running tools"
+        }"#,
+    );
+
+    assert_eq!(event.event_type, "tool_call");
+    let statuses = parsed
+        .into_iter()
+        .filter_map(|event| match event {
+            BackendEvent::Status(status) => Some(status),
+            _ => None,
+        })
+        .collect::<Vec<_>>();
+    assert_eq!(
+        statuses,
+        vec!["tool  exec", "tool  shell", "tool  write_file"]
+    );
+}
+
+#[test]
+fn contract_array_field_reports_shape_errors() {
+    let payload = json!({"list": {"not": "an array"}});
+    let err = expect_array_field(&payload, "list", "sessions.list").expect_err("should fail");
+    assert!(err.to_string().contains("sessions.list contract error"));
+}
+
+#[test]
+fn contract_builds_provider_verify_args_without_user_id() {
+    let mutation = crate::backend::ProviderMutation {
+        name: Some("demo".to_string()),
+        base_url: Some("https://example.com".to_string()),
+        api_key: None,
+        model: Some("demo-chat".to_string()),
+        is_default: Some(false),
+    };
+
+    let args = CliJsonCommand::ProviderVerify {
+        mutation: &mutation,
+    }
+    .args();
+    assert_eq!(
+        args,
+        vec![
+            "provider",
+            "verify",
+            "--json",
+            "--name",
+            "demo",
+            "--base-url",
+            "https://example.com",
+            "--model",
+            "demo-chat",
+            "--unset-default",
+        ]
+    );
+}
