+++ app/terminal/src/terminal/tests/mod.rs
@@ -1,0 +1,3 @@
+mod formatting;
+mod interaction;
+mod provider_args;
